requirements.txt
streamlit==1.35.0
snowflake-connector-python==3.14.0
pandas==2.2.2
rich==13.7.1
python-dotenv==1.0.1
